% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{calculateCentroids}
\alias{calculateCentroids}
\title{calculateCentroids}
\usage{
calculateCentroids(object,
  community.detection.algorithm = spinglass.community)
}
\arguments{
\item{object}{GNG object}

\item{community.detection.algorithm}{Used algorithm from igraph package, by default spinglass.community}
}
\description{
Using passed community.detection finds communities and for each community pick node with biggest betweenness score
}
\details{
Get centroids
}
\examples{
gng <- GNG(gng.preset.sphere(100))
print(node(gng, calculateCentroids(gng)[1])$pos)
}

