% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\docType{data}
\name{caret.gmumSvmLinear}
\alias{caret.gmumSvmLinear}
\title{Caret model representation for SVM with linear kernel}
\format{List of caret specific values}
\usage{
caret.gmumSvmLinear
}
\description{
Supply as parameter "method" in the caret::train function
}
\examples{
\dontrun{
model <- train(Class ~ ., data = training,
method = caret.gmumSvmLinear,
preProc = c("center", "scale"),
tuneLength = 8,
trControl = fitControl,
tuneGrid = expand.grid(C=10^(c(-4:4)), gamma=10^(c(-4:4))),
core = "libsvm", # gmum.R parameter - pick library
verbosity = 0 # no outputs
)
}
}
\keyword{datasets}

