% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{cec.ellipsegauss.extra}
\alias{cec.ellipsegauss.extra}
\title{cec.ellipsegauss.extra}
\description{
Extra information for dataset cec.ellipsegauss (energy and cluster assignment)
}
\keyword{data}

