% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{cec.mouse1.spherical.extra}
\alias{cec.mouse1.spherical.extra}
\title{cec.mouse1.spherical.extra}
\description{
Extra information for dataset cec.mouse1.extra (energy and cluster assignment)
}
\keyword{data}

