% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{centers.cec}
\alias{centers.cec}
\title{centers}
\format{\preformatted{ NULL
}}
\usage{
centers.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print centers of clusters
}
\keyword{datasets}

