% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{centers}
\alias{centers}
\alias{centers,Rcpp_CecModel-method}
\title{centers}
\usage{
centers(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print centers of clusters
}
\examples{
\dontrun{
centers(c)
}
}

