% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{centroids.gng}
\alias{centroids}
\alias{centroids.gng}
\title{centroids}
\usage{
centroids(gng)
}
\description{
Using infomap.communities finds communities and for each community pick node with biggest betweenness score
}
\details{
Get centroids
}
\examples{
# Print position of the first centroid
print(node(gng, centroids(gng)[1])$pos)
}

