% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{clustering.gng}
\alias{clustering}
\alias{clustering.gng}
\title{clustering}
\format{\preformatted{ NULL
}}
\usage{
clustering(gng)
}
\description{
Gets vector with node indexes assigned to examples in the dataset
}
\keyword{datasets}

