% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{clustering.cec}
\alias{clustering.cec}
\title{clustering}
\format{\preformatted{ NULL
}}
\usage{
clustering.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print labels assigned
}
\keyword{datasets}

