% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{convertToGraph.gng}
\alias{convertToGraph}
\alias{convertToGraph.gng}
\title{convertToGraph}
\usage{
convertToGraph(gng)
}
\arguments{
\item{gng}{Learned GNG object}
}
\description{
Converts GNG to igraph object.
}

