% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{convertToGraph}
\alias{convertToGraph}
\title{convertToGraph}
\usage{
convertToGraph(object, calculate.dist = TRUE)
}
\arguments{
\item{object}{GNG object}

\item{calculate.dist}{If true will calculate all \code{n^2} distances in the graph}
}
\description{
Converts GNG to igraph object.
}

