% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{convertToIGraph}
\alias{convertToIGraph}
\title{convertToIGraph}
\usage{
convertToIGraph(object, calculate.dist = TRUE)
}
\arguments{
\item{object}{GNG object}

\item{calculate.dist}{If true will calculate all \code{n^2} distances in the graph}
}
\description{
Converts GNG to igraph object, where every vertex contains attributes gng.index, error, data.label and 3 first spatial coordinates (as attributes v0, v1, v2).
Additionally utility attribute is present if utility GNG is used.
}

