% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{covMatrix.cec}
\alias{covMatrix.cec}
\title{covMatrix}
\format{\preformatted{ NULL
}}
\usage{
covMatrix.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print covariances of clusters
}
\keyword{datasets}

