% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{covMatrix}
\alias{covMatrix}
\alias{covMatrix,Rcpp_CecModel-method}
\title{covMatrix}
\format{\preformatted{ NULL
}}
\usage{
covMatrix
}
\description{
Print covariances of clusters
}
\examples{
\dontrun{
covMatrix(c)
}
}
\keyword{datasets}

