% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{energy.cec}
\alias{energy.cec}
\title{energy}
\format{\preformatted{ NULL
}}
\usage{
energy.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print result energy of clustering found
}
\keyword{datasets}

