% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{energy}
\alias{energy}
\alias{energy,Rcpp_CecModel-method}
\title{energy}
\format{\preformatted{ NULL
}}
\usage{
energy
}
\description{
Print result energy of clustering found
}
\keyword{datasets}

