% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{errorStatistics}
\alias{errorStatistics}
\title{errorStatistics}
\usage{
errorStatistics(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Gets vector with errors for every second of execution
}
\examples{
gng <- GNG(gng.preset.sphere(100))
errorStatistics(gng)
}

