% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.utils.R
\name{get.wine.dataset.X}
\alias{get.wine.dataset.X}
\title{get.wine.dataset.X}
\usage{
get.wine.dataset.X(scale = TRUE)
}
\arguments{
\item{scale}{if TRUE will perform feature scaling}
}
\description{
Retrieves wine dataset design matrix from UCI repository
}

