% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{getDataset}
\alias{getDataset}
\alias{getDataset,Rcpp_CecModel-method}
\title{getDataset}
\usage{
getDataset(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print input dataset
}
\examples{
\dontrun{
getDataset(c)
}
}

