% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{gng.plot.layout.v2d}
\alias{gng.plot.layout.v2d}
\title{Use first two spatial coordinates as position in layout}
\usage{
gng.plot.layout.v2d(g)
}
\arguments{
\item{g}{GNG object}
}
\description{
Use first two spatial coordinates as position in layout
}
\note{
You can pass any igraph layout algorithm to plot
}

