% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.presets.R
\name{gng.preset.cube}
\alias{gng.preset.cube}
\title{gng.preset.cube}
\usage{
gng.preset.cube(N, r = 0.5, center = c(0.5, 0.5, 0.5))
}
\arguments{
\item{N}{Number of points}

\item{r}{Length of the side of cube}

\item{center}{Center of the plane}
}
\description{
Generate sample cube dataset
}
\details{
Cube preset dataset
}
\examples{
X <- gng.preset.cube(100)
gng <- GNG(X)
}

