% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.presets.R
\name{gng.preset.plane}
\alias{gng.preset.plane}
\title{gng.preset.plane}
\usage{
gng.preset.plane(N, side = 0.5, center = c(0.5, 0.5, 0.5))
}
\arguments{
\item{N}{Number of points}

\item{side}{Length of the side of plane}

\item{center}{Center of the plane}
}
\description{
Generate sample plane dataset
}
\details{
Plane preset dataset
}
\examples{
X <- gng.preset.plane(100)
gng <- GNG(X)
}

