% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.presets.R
\name{gng.preset.sphere}
\alias{gng.preset.sphere}
\title{gng.preset.sphere}
\usage{
gng.preset.sphere(N, r = 0.5, center = c(0.5, 0.5, 0.5))
}
\arguments{
\item{N}{Number of points}

\item{r}{Radius of the sphere}

\item{center}{Center of the sphere}
}
\description{
Generate sample sphere dataset
}
\details{
Sphere preset dataset
}
\examples{
X <- gng.preset.sphere(100)
gng <- GNG(X)
}

