% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{gngSave}
\alias{gngSave}
\title{gngSave}
\usage{
gngSave(gng)
}
\arguments{
\item{filename}{File where binary will be saved}
}
\description{
Writes model to a disk space efficient binary format.
}
\details{
Save model to binary format
}
\examples{
gngSave(gng, 'graph.bin')
}

