% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{gngSave}
\alias{gngSave}
\title{gngSave}
\usage{
gngSave(object, filename)
}
\arguments{
\item{object}{GNG object}

\item{filename}{File where binary will be saved}
}
\description{
Writes model to a disk space efficient binary format.
}
\details{
Save model to binary format
}

