% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{insertExamples.gng}
\alias{insertExamples}
\alias{insertExamples.gng}
\title{insertExamples}
\usage{
insertExamples(gng, M, L=c())
}
\arguments{
\item{examples}{\code{matrix} or \code{data.frame} with rows as examples. Note: if training online make sure
number of columns matches dim parameter passed to GNG constructor.}

\item{labels}{\code{vector} of labels, that will be associated with nodes in the graph. GNG will assign to each
node a mean of labels of closest examples.}
}
\description{
Insert examples with optional labels.
}
\examples{
#Add preset examples
M = generateExamples(preset=gng.preset.sphere)
insertExamples(gng, M)
}

