% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{insertExamples}
\alias{insertExamples}
\title{insertExamples}
\usage{
insertExamples(object, examples, labels = c())
}
\arguments{
\item{object}{GNG object}

\item{examples}{\code{matrix} or \code{data.frame} with rows as examples. Note: if training online make sure
number of columns matches dim parameter passed to GNG constructor.}

\item{labels}{\code{vector} of labels, that will be associated with nodes in the graph. GNG will assign to each
node a mean of labels of closest examples.}
}
\description{
Insert examples with optional labels.
}
\note{
It copies your examples twice in RAM. You might want to use object$insertExamples, or
not to copy at all set_memory_move_examples (when using this function, remember not to modify the matrix
and after removing the object delete it aswell)
}
\examples{
\dontrun{
data(wine, package="rattle")
scaled.wine <- scale(wine[-1])
gng <- GNG(scaled.wine)
# Add preset examples
M = generateExamples(preset=gng.preset.sphere)
insertExamples(gng, M)
}
}

