% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{isRunning}
\alias{isRunning}
\title{isRunning}
\usage{
isRunning(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Returns TRUE if GNG object is training
}
\details{
Check if GNG is running
}
\examples{
gng <- GNG(gng.preset.sphere(100))
# FALSE, because did not pass train.online to constructor
print(isRunning(gng))
}

