% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{iterations}
\alias{iterations}
\alias{iterations,Rcpp_CecModel-method}
\title{iterations}
\usage{
iterations(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print how many iterations it took to learn CEC model
}
\examples{
\dontrun{
iterations(c)
}
}

