% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{load.gng}
\alias{load.gng}
\title{load.gng}
\usage{
load.gng(gng)
}
\arguments{
\item{filename}{Dump location}
}
\description{
Writes model to a disk space efficient binary format.
}
\details{
Load model from binary format
}
\examples{
load.gng('model.bin')
}

