% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{log.energy.cec}
\alias{log.energy.cec}
\title{log.energy.cec}
\usage{
\method{log}{energy.cec}(c)
}
\arguments{
\item{c}{CEC model object}
}
\description{
Print energy recorded that has been recorded at each stage of learning.
Data is recorded only if you has chosen to when you created CEC model object.
}

