% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{log.iters.cec}
\alias{log.iters.cec}
\title{log.iters.cec}
\usage{
\method{log}{iters.cec}(c)
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print how many iterations it took to learn CEC model
}

