% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{log.ncluster.cec}
\alias{log.ncluster.cec}
\title{log.ncluster.cec}
\usage{
\method{log}{ncluster.cec}(c)
}
\arguments{
\item{c}{CEC model object}
}
\description{
Print number of clusters that has been recorded at each stage of learning.
Data is recorded only if you has chosen to when you created CEC model object.
}

