% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{logEnergy}
\alias{logEnergy}
\alias{logEnergy,Rcpp_CecModel-method}
\title{logEnergy}
\usage{
logEnergy(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print energy that has been recorded at each stage of learning.
Data is recorded only if you have chosen to when you created CEC model object.
}

