% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{meanError.gng}
\alias{meanError}
\alias{meanError.gng}
\title{meanError}
\usage{
meanError(gng)
}
\description{
Gets mean error of the graph (note: blocks the execution, O(n))
}
\examples{
meanError(gng)
}

