% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{meanError}
\alias{meanError}
\title{meanError}
\usage{
meanError(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Gets mean error of the graph (note: blocks the execution, O(n))
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
meanError(gng)
}
}

