% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{node}
\alias{node}
\title{node}
\usage{
node(x, gng_id)
}
\arguments{
\item{x}{GNG object}

\item{gng_id}{Id of the node to retrieve. This is the id returned by functions like predict, or centroids}
}
\description{
Retrieves node from resulting graph
}
\details{
Get GNG node
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
print(node(gng, 10)$pos)
}
}

