% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{nstart.cec}
\alias{nstart.cec}
\title{nstart}
\format{\preformatted{ NULL
}}
\usage{
nstart.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print number of times CEC algorithm is to be performed. The best outcome
is chosen as result.
}
\keyword{datasets}

