% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{pause}
\alias{pause}
\title{pause}
\usage{
pause(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Pause algorithm
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
pause(gng)
print(gng$isRunning())
}
}

