% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{plot}
\name{plot.svm}
\alias{plot.svm}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
plot(svm)
}
\description{
Draws choosen dimenstions from a dataset on 2D plot. NOTE: This function will be change to a package default function.
}
\keyword{datasets}

