% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.plot.R, R/svm.R
\docType{data}
\name{plot.Rcpp_CecModel}
\alias{plot.Rcpp_CecModel}
\alias{plot.Rcpp_SVMClient}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
\method{plot}{Rcpp_CecModel}(x, slice = c(), pca = FALSE,
  ellipses = FALSE, centers = FALSE, ...)

plot.Rcpp_SVMClient
}
\arguments{
\item{x}{CEC model object.}

\item{slice}{List of dimentions chosen for display since plot is 2D.}

\item{pca}{Apply PCA or not}

\item{ellipses}{Outline clusters.}

\item{centers}{Marks center of every cluster.}

\item{...}{other arguments not used by this method.}

\item{x}{Trained SVM object}

\item{X}{Optional new data points to be predicted and plotted in one of the following formats:
\code{data.frame}, \code{data.matrix}; default: \code{NULL}}

\item{mode}{Which plotting mode to use as string, available are:
\itemize{
\item \code{'normal'} - default mode, plots data in cols argument and a linear decision
  boundry in available
\item \code{'pca'} - preforms PCA decomposition and draws data in a subspace of first 2 dimensions
from the PCA
\item \code{'contour'} - countour plot for non-linear kernels
}}

\item{cols}{Data dimensions to be plotted as vector of length 2, default: \code{c(1,2)}}

\item{radius}{Radius of the plotted data points as float, default: \code{3}}

\item{radius.max}{Maximum radius of data points can be plotted, when model is trained
with example weights as float, default: \code{10}}

\item{...}{other arguments not used by this method.}
}
\description{
Plot clustering found on 2D plot coloring by cluster.

Plots trained svm data and models disciriminative
}
\details{
Plot CEC
}
\examples{
\dontrun{
plot(cec)
plot(cec, slice=c(1,3), ellipses=TRUE)
plot(cec, slice=c(1,2,3))
plot(cec, ellipses=TRUE, centers=FALSE)
plot(cec, pca=TRUE, ellipses=TRUE, centers=FALSE)
}
\dontrun{
plot(svm)
plot(svm, X=x, cols=c(1,3))
plot(svm, mode="pca", radius=5)
}
}
\keyword{datasets}

