% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.plot.R
\name{plot.Rcpp_CecModel}
\alias{plot.Rcpp_CecModel}
\title{plot}
\usage{
\method{plot}{Rcpp_CecModel}(x, slice = c(), pca = FALSE,
  ellipses = FALSE, centers = FALSE, ...)
}
\arguments{
\item{x}{CEC model object.}

\item{slice}{List of dimentions chosen for display since plot is 2D.}

\item{pca}{Apply PCA or not}

\item{ellipses}{Outline clusters.}

\item{centers}{Marks center of every cluster.}

\item{...}{other arguments not used by this method.}
}
\description{
Plot clustering found on 2D plot coloring by cluster.
}
\details{
Plot CEC
}
\examples{
\dontrun{
plot(cec)
plot(cec, slice=c(1,3), ellipses=TRUE)
plot(cec, slice=c(1,2,3))
plot(cec, ellipses=TRUE, centers=FALSE)
plot(cec, pca=TRUE, ellipses=TRUE, centers=FALSE)
}
}

