% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R, R/gng.R, R/svm.R
\docType{data}
\name{predict}
\alias{predict}
\alias{predict,Rcpp_CecModel-method}
\alias{predict.Rcpp_CecModel}
\alias{predict.Rcpp_GNGServer}
\alias{predict.Rcpp_SVMClient}
\title{predict}
\format{\preformatted{ NULL
}}
\usage{
\method{predict}{Rcpp_CecModel}(object, x, ...)

\method{predict}{Rcpp_GNGServer}(object, x, ...)

predict.Rcpp_SVMClient
}
\arguments{
\item{object}{Trained CEC model object.}

\item{x}{Given point.}

\item{...}{other arguments not used by this method.}

\item{gng}{GNG object}

\item{x}{Vector or matrix of examples}

\item{object}{Trained SVM object}

\item{x_test}{Unlabeled data, in one of the following formats:
\code{data.frame}, \code{data.matrix}, \code{SparseM::matrix.csr}, \code{Matrix::Matrix},
\code{slam::simple_triplet_matrix}}

\item{decision.function}{Uf \code{TRUE} returns SVMs decision function
(distance of a point from discriminant) instead of predicted labels, default: \code{FALSE}}

\item{...}{other arguments not used by this method.}
}
\description{
Classify a new point according to the model (returns index of cluster where given point belong to)

Retrieves prediction from GNG

Returns predicted classes or distance to discriminative for provided test examples.
}
\details{
Predict
}
\examples{
\dontrun{
predict(gng, c(1,2,2))
}
\dontrun{
svm <- SVM(x, y, core="libsvm", kernel="linear", C=1)
predcit(svm, x_test)
}
}
\keyword{datasets}

