% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{predict.cec}
\alias{predict.Rcpp_CecModel}
\alias{predict.Rcpp_CecModel-method}
\alias{predict.cec}
\title{predict}
\usage{
\method{predict}{Rcpp_CecModel}(object, x, ...)
}
\arguments{
\item{object}{Trained CEC model object.}

\item{x}{Given point.}

\item{...}{other arguments not used by this method.}
}
\description{
Classify a new point according to the model (returns index of cluster where given point belong to)
}

