% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{predict.gng}
\alias{predict.Rcpp_GNGServer}
\alias{predict.gng}
\title{predict}
\usage{
\method{predict}{Rcpp_GNGServer}(object, x, ...)
}
\arguments{
\item{object}{Trained model}

\item{x}{Vector or matrix of examples}

\item{...}{other arguments not used by this method}
}
\description{
Retrieves prediction from trained GNG model
}
\details{
Predict
}
\examples{
\dontrun{
predict(gng, c(1,2,2))
}
}

