% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\docType{methods}
\name{predict.svm.gmum}
\alias{predict.svm.gmum}
\title{Predict}
\usage{
predict(svm, x)
}
\arguments{
\item{object}{Trained SVM object.}

\item{x}{unlabeled data, in one of the following formats:
\code{data.frame}, \code{data.matrix}, \code{SparseM::matrix.csr}, \code{Matrix::Matrix},
\code{slam::simple_triplet_matrix}}

\item{decision.function}{if \code{TRUE} returns SVMs decision function
(distance of a point from discriminant) instead of predicted labels, default: \code{FALSE}}
}
\description{
Returns predicted classes or distance to discriminative for provided test examples.
}

