% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{predictCentroid}
\alias{predictCentroid}
\title{predictCentroid}
\usage{
predictCentroid(gng, centroids(gng), c(1,1,1))
}
\arguments{
\item{centroids}{List of indexes of nodes in gng.
Will assign each point to one of indexes from centroids.
Can be found using centroids.gng function.}

\item{x}{Can be either vector or data.frame.}
}
\description{
Finds closest centroid from given list
}
\details{
Find closest centroid
}
\examples{
# Find closest centroid to c(1,1,1)
found.centroids <- centroids(gng)
predictCentroid(gng, found.centroids, c(1,1,1))
}

