% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{predictCentroids}
\alias{predictCentroids}
\title{predictCentroids}
\usage{
predictCentroids(object, community.detection.algorithm = spinglass.community)
}
\arguments{
\item{object}{GNG object}

\item{community.detection.algorithm}{# STASZEK PLZ DOCUMENT THIS}
}
\description{
Using infomap.communities finds communities and for each community pick node with biggest betweenness score
}
\details{
Get centroids
}
\examples{
\dontrun{
print(node(gng, predictCentroids(gng)[1])$pos)
}
}

