% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predictCluster.cec}
\alias{predictCluster.cec}
\title{predictCluster}
\usage{
predictCluster.cec(c, vec)
}
\arguments{
\item{c}{CEC model object}

\item{vec}{given point}
}
\description{
For given point predict cluster where it belongs
}

