% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predictClusters.cec}
\alias{predictClusters.cec}
\title{predictClusters}
\usage{
predictClusters.cec(c, vec)
}
\arguments{
\item{c}{CEC model object}

\item{vec}{given point}
}
\description{
For given point print propabilities of belonging to each cluster
}

