% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.print.R, R/gng.R, R/svm.R
\docType{methods}
\name{print.Rcpp_CecModel}
\alias{print.Rcpp_CecModel}
\alias{print.Rcpp_GNGServer}
\alias{print.Rcpp_SVMClient}
\title{print}
\format{\preformatted{ NULL
}}
\usage{
\method{print}{Rcpp_CecModel}(x, ...)

\method{print}{Rcpp_GNGServer}(x, ...)

print.Rcpp_SVMClient
}
\arguments{
\item{x}{CEC object model.}

\item{...}{other arguments not used by this method.}

\item{x}{GNG object model.}

\item{...}{other arguments not used by this method.}

\item{object}{Trained SVM object}

\item{...}{other arguments not used by this method.}
}
\description{
Print basic information about clusters found.
Presents a structure of the cec results object (clusters found)

Print basic information about GNG object

Prints short summary of a trained model.
}
\details{
Print CEC

Print CEC
}
\keyword{datasets}

