% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dataset.X}
\alias{dataset}
\alias{dataset.X}
\title{dataset.X}
\format{\preformatted{ NULL
}}
\usage{
dataset.X(svm)
}
\arguments{
\item{object}{SVM object.}
}
\description{
Prints dataset stored in a SVM object, without the labels.
}
\keyword{datasets}

