% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print.cec}
\alias{print.cec}
\title{print}
\usage{
print(cec)
}
\arguments{
\item{x}{CEC object model}
}
\description{
Print basic information about clusters found.
Presents a structure of the cec results object (clusters found)
}
\details{
Print
}

