% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.utils.R
\docType{methods}
\name{print.Rcpp_GNGServer}
\alias{print.Rcpp_GNGServer}
\title{print}
\usage{
\method{print}{Rcpp_GNGServer}(x, ...)
}
\arguments{
\item{x}{GNG object model.}

\item{...}{other arguments not used by this method.}
}
\description{
Print basic information about GNG object
}

