% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\docType{methods}
\name{print.svm}
\alias{print.svm}
\title{print}
\format{\preformatted{ NULL
}}
\usage{
print(svm)
}
\arguments{
\item{object}{SVM object}
}
\description{
Prints short summary of the SVM object and its parameters.
}
\keyword{datasets}

