% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{class}
\name{Rcpp_CecConfiguration-class}
\alias{Rcpp_CecConfiguration-class}
\title{Class Rcpp_CecConfiguration.}
\description{
Class \code{Rcpp_CecConfiguration} defines a CEC model configuration class.
}

