% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{class}
\name{Rcpp_GNGServer-class}
\alias{Rcpp_GNGServer-class}
\title{Class Rcpp_GNGServer.}
\description{
Class \code{Rcpp_GNGServer} defines a GNGServer class.
}
\section{Methods}{

\describe{
\item{\code{getClustering()}}{Rcpp::NumericVector getClustering()}

\item{\code{getCurrentIteration()}}{unsigned int getCurrentIteration()  const }

\item{\code{getDatasetSize()}}{unsigned int getDatasetSize()  const }

\item{\code{getErrorStatistics()}}{Rcpp::NumericVector getErrorStatistics()}

\item{\code{getMeanError()}}{double getMeanError()}

\item{\code{getNode(...)}}{Rcpp::List getNode(int)}

\item{\code{getNumberNodes()}}{unsigned int getNumberNodes()  const }
}}

