% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{run}
\alias{run}
\title{run}
\usage{
run(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Run algorithm (in parallel)
}
\examples{
gng <- GNG(gng.preset.sphere(100))
run(gng)
print(isRunning(gng))
}

