% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{runAll.cec}
\alias{runAll.cec}
\title{runAll}
\format{\preformatted{ NULL
}}
\usage{
runAll.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Starts whole algorithm again with same parameters
}
\keyword{datasets}

