% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{runAll}
\alias{runAll}
\alias{runAll,Rcpp_CecModel-method}
\title{runAll}
\format{\preformatted{ NULL
}}
\usage{
runAll
}
\description{
Starts whole algorithm again with same parameters
}
\keyword{datasets}

