% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{runOneIteration.cec}
\alias{runOneIteration.cec}
\title{runOneIteration}
\usage{
runOneIteration.cec(c)
}
\arguments{
\item{c}{CEC model object.}
}
\description{
runs one iteration of algorithm
}

