% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{runOneIteration}
\alias{runOneIteration}
\alias{runOneIteration,Rcpp_CecModel-method}
\title{runOneIteration}
\usage{
runOneIteration(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
runs one iteration of algorithm
}
\examples{
\dontrun{
runOneIteration(c)
}
}

