% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.summary.R, R/gng.R, R/svm.R
\docType{methods}
\name{summary.Rcpp_CecModel}
\alias{summary.Rcpp_CecModel}
\alias{summary.Rcpp_GNGServer}
\alias{summary.Rcpp_SVMClient}
\title{summary}
\format{\preformatted{ NULL
}}
\usage{
\method{summary}{Rcpp_CecModel}(object, ...)

\method{summary}{Rcpp_GNGServer}(object, ...)

summary.Rcpp_SVMClient
}
\arguments{
\item{object}{CEC model object.}

\item{...}{other arguments not used by this method.}

\item{x}{GNG object model.}

\item{...}{other arguments not used by this method.}

\item{object}{Trained SVM object}

\item{...}{other arguments not used by this method.}
}
\description{
Print detailed information about CEC model object

Print basic information about GNG object

Prints short summary of a trained model.
}
\details{
Summary CEC

Summary of CEC
}
\keyword{datasets}

