% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.utils.R
\docType{methods}
\name{summary.Rcpp_GNGServer}
\alias{summary.Rcpp_GNGServer}
\title{summary}
\usage{
\method{summary}{Rcpp_GNGServer}(object, ...)
}
\arguments{
\item{object}{GNG object model.}

\item{...}{other arguments not used by this method.}
}
\description{
Print basic information about GNG object
}
\details{
Summary of GNG object
}

