% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\name{summary.Rcpp_SVMClient}
\alias{summary.Rcpp_SVMClient}
\title{Summary of SVM object}
\usage{
\method{summary}{Rcpp_SVMClient}(object, ...)
}
\arguments{
\item{object}{Trained SVM object}

\item{...}{other arguments not used by this method.}
}
\description{
Prints short summary of a trained model.
}

