% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{load.dataset}
\alias{dataset}
\alias{load.dataset}
\title{load.dataset}
\format{\preformatted{ NULL
}}
\usage{
load.dataset(svm, x)
}
\arguments{
\item{object}{SVM object to which examples will be loaded.}

\item{x}{Labeled dataset with lables in the last column.}
}
\description{
Loads provided dataset to a SVM object, right now lables need to be in the last column.
}
\keyword{datasets}

