% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm}
\alias{parameters}
\alias{params.svm}
\title{params}
\format{\preformatted{ NULL
}}
\usage{
params(svm, object, lib, kernel, preprocess, C, gamma, coef0, degree, cache_size, shrinking, probability)
}
\arguments{
\item{lib}{Desired SVM Library, avialable are: libsvm.}

\item{kernel}{Kernel type, avialable are: linear, poly, rbf, sigmoid.}

\item{prep}{Preprocess method, avialable are: none, 2e.}

\item{C}{Cost/Complexity parameter.}

\item{gamma}{Gamma parameter for poly, rbf and sigmoid kernels.}

\item{coef0}{Coef0 for poly and sigmoid kernels.}

\item{degree}{Degree for poly kernel.}

\item{shrinking}{Whether to use shrinking heuristics.}

\item{probability}{Whether to train a model for probability estimates .}

\item{cache_size}{Cache size.}
}
\description{
Changes SVM objects parameters to provided, not every parameter needs to be given.
}
\examples{
params(svm, kernel="linear", preproces="none", C=10000)
}
\keyword{datasets}

