% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.C}
\alias{parameters}
\alias{params.svm.C}
\title{params.C}
\format{\preformatted{ NULL
}}
\usage{
params.C(svm)
params.C(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New cost.}
}
\description{
Changes or prints (if a argument is not provided) Cost parameter in a SVM object, note: requires retraining. Aviable are: none, 2e
}
\examples{
params.C(svm, 1000)
}
\keyword{datasets}

