% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.cache_size}
\alias{parameters}
\alias{params.svm.cache_size}
\title{params.cache_size}
\format{\preformatted{ NULL
}}
\usage{
params.cache_size(svm)
params.cache_size(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New cache size.}
}
\description{
Changes or prints (if a argument is not provided) cache_size (in MB) in a SVM object, note: requires retraining.
}
\examples{
params.cache_size(svm, 100)
}
\keyword{datasets}

