% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.coef0}
\alias{parameters}
\alias{params.svm.coef0}
\title{params.coef0}
\format{\preformatted{ NULL
}}
\usage{
params.coef0(svm)
params.coef0(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{coef0.}
}
\description{
Changes or prints (if a argument is not provided) coef0 parameter in a SVM object, note: requires retraining.
}
\examples{
params.coef0(svm, 0)
}
\keyword{datasets}

