% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.degree}
\alias{parameters}
\alias{params.svm.degree}
\title{params.degree}
\format{\preformatted{ NULL
}}
\usage{
params.degree(svm)
params.degree(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New degree.}
}
\description{
Changes or prints (if a argument is not provided) degree in a SVM object, note: requires retraining.
}
\examples{
params.degree(svm, 4)
}
\keyword{datasets}

