% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.gamma}
\alias{parameters}
\alias{params.svm.gamma}
\title{params.gamma}
\format{\preformatted{ NULL
}}
\usage{
params.gamma(svm)
params.gamma(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New gamma.}
}
\description{
Changes or prints (if a argument is not provided) gamma parameter in a SVM object, note: requires retraining.
}
\examples{
params.gamma(svm, 0.01)
}
\keyword{datasets}

