% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.kernel}
\alias{parameters}
\alias{params.svm.kernel}
\title{params.kernel}
\format{\preformatted{ NULL
}}
\usage{
params.kernel(svm)
params.kernel(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New kernel.}
}
\description{
Changes or prints (if a argument is not provided)kernel in a SVM object, note: requires retraining. Aviable are: libsvm.
}
\examples{
params.kernel(svm, "rbf")
}
\keyword{datasets}

