% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.lib}
\alias{parameters}
\alias{params.svm.lib}
\title{params.lib}
\format{\preformatted{ NULL
}}
\usage{
params.lib(svm)
params.lib(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New library.}
}
\description{
Changes or prints (if a argument is not provided) library in a SVM object, note: requires retraining.
}
\examples{
params.lib(svm, "libsvm")
}
\keyword{datasets}

