% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.prep}
\alias{parameters}
\alias{params.svm.prep}
\title{params.prep}
\format{\preformatted{ NULL
}}
\usage{
params.prep(svm)
params.prep(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{New preprocess type.}
}
\description{
Changes or prints (if a argument is not provided) preprocess type in a SVM object, note: requires retraining. Aviable are: linear, sigmoid, poly, rbf.
}
\examples{
params.prep(svm, "2e")
}
\keyword{datasets}

