% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.probability}
\alias{parameters}
\alias{params.svm.probability}
\title{params.probability}
\format{\preformatted{ NULL
}}
\usage{
params.probability(svm)
params.probability(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{TRUE/FALSE.}
}
\description{
Changes or prints (if a argument is not provided) if a SVM object is used for probability estimates , note: requires retraining.
}
\examples{
params.probability(svm, TRUE)
}
\keyword{datasets}

