% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{params.svm.shrinking}
\alias{parameters}
\alias{params.svm.shrinking}
\title{params.shrinking}
\format{\preformatted{ NULL
}}
\usage{
params.shrinking(svm)
params.shrinking(svm,x)
}
\arguments{
\item{object}{SVM object.}

\item{x}{TRUE/FALSE.}
}
\description{
Changes or prints (if a argument is not provided) if shrinking heurestics are used in a SVM object, note: requires retraining.
}
\examples{
params.shrinking(svm, TRUE)
}
\keyword{datasets}

