% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict.svm}
\alias{predict.svm}
\alias{test}
\title{Predict}
\format{\preformatted{ NULL
}}
\usage{
predict(svm, x)
}
\arguments{
\item{object}{Trained SVM object.}

\item{x}{unlabeled data, note that each entry needs to be the same dimentionality as training examples.}
}
\description{
Returns predicted classes for provided test examples, note that provided model needs to be trained first before any prediction can be made.
}
\keyword{datasets}

