% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print.svm}
\alias{print.svm}
\alias{summary}
\title{print}
\format{\preformatted{ NULL
}}
\usage{
print(svm)
}
\arguments{
\item{object}{SVM object}
}
\description{
Prints short summary of the SVM object and its parameters.
}
\keyword{datasets}

