% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{train.svm}
\alias{train}
\alias{train.svm}
\title{Train SVM model}
\format{\preformatted{ NULL
}}
\usage{
train(svm)
}
\arguments{
\item{object}{Trained SVM object which will be used for prediction.}
}
\description{
Trains provided svm object, using its parameters and dataset.
}
\keyword{datasets}

