% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.utils.R
\name{svm.accuracy}
\alias{svm.accuracy}
\title{Measure accuracy scoreof a prediction}
\usage{
svm.accuracy(prediction, target)
}
\arguments{
\item{prediction}{factor or 1 dim vector with predicted classes}

\item{target}{factor or 1 dim vector with true classes

#'}
}
\description{
Calculates accuracy of a prediction, returns precent of correctly predicted examples
over all test examples.
}
\examples{
\dontrun{
# firstly, SVM model needs to be trained
svm <- SVM(x, y, core="libsvm", kernel="linear", C=1)
# then we can use it to predict unknown samples
p <- predcit(svm, x_test)
acc <- svm.accuracy(p, y)
}
}

