% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\docType{data}
\name{caret.gmumSvmLinear}
\alias{caret.gmumSvmLinear}
\title{Caret model representation for SVM with linear kernel}
\format{\preformatted{List of 11
 $ label     : chr "gmum.r.svmRadial"
 $ library   : chr "gmum.r"
 $ type      : chr "Classification"
 $ parameters:'data.frame':	1 obs. of  3 variables:
  ..$ parameter: Factor w/ 1 level "C": 1
  ..$ class    : Factor w/ 1 level "double": 1
  ..$ label    : Factor w/ 1 level "C": 1
 $ grid      :function (x, y, len = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1135 33 1137 1 33 1 1135 1137
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
 $ fit       :function (x, y, wts, param, lev, last, classProbs, ...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1075 9 1111 3 9 3 1075 1111
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
 $ predict   :function (modelFit, newdata, submodels = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1112 13 1114 3 13 3 1112 1114
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
 $ prob      :function (modelFit, newdata, submodels = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1115 10 1117 3 10 3 1115 1117
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
 $ varImp    : NULL
 $ levels    :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1119 12 1121 3 12 3 1119 1121
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
 $ sort      :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1122 10 1123 20 10 20 1122 1123
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10fd4c4a0> 
}}
\usage{
caret.gmumSvmLinear
}
\description{
Supply as parameter "method" in the caret::train function
}
\keyword{datasets}

