% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{CEC}
\alias{CEC}
\title{CEC()}
\usage{
CEC(x = NULL, k = 0, method.type = "", method.init = "kmeans++",
  params.r = 0, params.cov = matrix(0), params.centroids = NULL,
  params.mix = NULL, params.function = "", control.nstart = 1,
  control.eps = 1e-04, control.itmax = 1, log.energy = FALSE,
  log.ncluster = FALSE, log.iters = FALSE)
}
\arguments{
\item{x}{numeric matrix of data}

\item{k}{initial number of clusters}

\item{method.type}{Gauss family}

\item{method.init}{method to initialize clusters}

\item{params.r}{radius for spherical family}

\item{params.cov}{covariance matrix for covariance family}

\item{params.centroids}{list of centroids}

\item{control.nstart}{how many times to perform algorithm}

\item{control.eps}{what change of value should terminate algorithm}

\item{control.itmax}{maximum number of iterations at each start}

\item{log.energy}{record collected energy of all clusters in each iteration}

\item{log.ncluster}{record number of clusters in each iteration}

\item{log.iters}{record number of iterations}
}
\description{
Create CEC model object
}
\details{
CEC
}

