% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{CEC}
\alias{CEC}
\title{CEC}
\usage{
CEC(k=3, x=dataset)
}
\arguments{
\item{x}{Numeric matrix of data.}

\item{k}{Initial number of clusters.}

\item{method.type}{Type of clustering (Gauss family).
\enumerate{
\item diagonal Gaussians with diagonal covariance. The clustering will try to divide the data into ellipsoid with radiuses parallel to coordinate axes
\item fsphere Spherical (radial) Gaussian densities (additional parameter - radius)
\item full The clustering will have the tendency to divide the data into clusters resembling the unit circles in the Mahalanobis distance (additional parameter - covaraince matrix required)
\item func Own function dependent on m and sigma (additional parameter)
\item mix Mix of others Gaussian types.
\item standard We divide dataset into ellipsoid-like clusters without any preferences (default)
\item sphere The clustering will try to divide the data into circles of arbitrary sizes}}

\item{method.init}{Method to initialize clusters.
\enumerate{
\item centroids
\item kmeans++
\item random}}

\item{params.r}{Radius for spherical family.}

\item{params.cov}{Covariance matrix for covariance family.}

\item{params.centroids}{List of centroids.}

\item{control.nstart}{How many times to perform algorithm.}

\item{control.eps}{What change of value should terminate algorithm.}

\item{control.itmax}{Maximum number of iterations at each start.}

\item{log.energy}{Records collected energy of all clusters in each iteration.}

\item{log.ncluster}{Records number of clusters in each iteration.}

\item{log.iters}{Records number of iterations.}
}
\description{
Create CEC model object
}
\details{
CEC
}

