% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{predict.cec}
\alias{predict.cec}
\title{predict}
\format{\preformatted{ NULL
}}
\usage{
predict.cec
}
\arguments{
\item{object}{Trained CEC model object.}

\item{x}{Given point.}
}
\description{
Classify a new point according to the model (returns index of cluster where given point belong to)
}
\keyword{datasets}

