\docType{methods}
\name{predictCluster.cec}
\alias{predictCluster.cec}
\alias{predictClusters.cec}
\title{predictCluster}
\format{\preformatted{ NULL
}}
\usage{
predictCluster.cec

predictClusters.cec
}
\arguments{
  \item{c}{Trained CEC model object.}

  \item{vec}{Given point.}

  \item{c}{Trained CEC model object.}

  \item{vec}{Given point.}
}
\description{
Returns cluster where belong given point.

Returns propabilities of belonging to each cluster for
given point.
}
\keyword{datasets}

