% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{centers}
\alias{centers}
\alias{centers,Rcpp_CecModel-method}
\title{centers}
\format{\preformatted{ NULL
}}
\usage{
centers
}
\description{
Print centers of clusters
}
\examples{
\dontrun{
centers(c)
}
}
\keyword{datasets}

