% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R, R/gng.R
\docType{methods}
\name{clustering}
\alias{clustering}
\alias{clustering.Rcpp_CecModel}
\alias{clustering.Rcpp_GNGServer}
\title{clustering}
\usage{
clustering(c)

\method{clustering}{Rcpp_CecModel}(c)

\method{clustering}{Rcpp_GNGServer}(c)
}
\arguments{
\item{c}{Object with clusters}
}
\description{
Print labels assigned

Print labels assigned

Gets vector with node indexes assigned to examples in the dataset
}
\examples{
\dontrun{
clustering(c)
}
\dontrun{
clustering(c)
}
\dontrun{
clustering(gng)
}
}

