% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{covMatrix}
\alias{covMatrix}
\alias{covMatrix,Rcpp_CecModel-method}
\title{covMatrix}
\usage{
covMatrix(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print covariances of clusters
}
\examples{
\dontrun{
covMatrix(c)
}
}

