% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{energy}
\alias{energy}
\alias{energy,Rcpp_CecModel-method}
\title{energy}
\usage{
energy(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print result energy of clustering found
}
\examples{
\dontrun{
energy(c)
}
}

