% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{findClosests}
\alias{findClosests}
\title{findClosests}
\usage{
findClosests(object, node.ids, x)
}
\arguments{
\item{object}{GNG object}

\item{node.ids}{List of indexes of nodes in gng.}

\item{x}{Can be either \code{vector} or \code{data.frame.}}
}
\description{
Finds closest node from given list to vector. Often used together with calculateCentroids
}
\details{
Find closest node
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
# Find closest centroid to c(1,1,1)
found.centroids <- calculateCentroids(gng)
findClosest(gng, found.centroids, c(1,1,1))
}
}

