% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{gngLoad}
\alias{gngLoad}
\title{gngLoad}
\usage{
gngLoad(gng)
}
\arguments{
\item{filename}{Dump location}
}
\description{
Writes model to a disk space efficient binary format.
}
\details{
Load model from binary format
}
\examples{
gngLoad('model.bin')
}

