% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{log.energy.cec}
\alias{log.energy.cec}
\title{log.energy.cec}
\usage{
\method{log}{energy.cec}(c)
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print energy that has been recorded at each stage of learning.
Data is recorded only if you have chosen to when you created CEC model object.
}

