% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{log.energy.cec}
\alias{log.energy.cec}
\title{log.energy.cec}
\format{\preformatted{ NULL
}}
\usage{
log.energy.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print energy that has been recorded at each stage of learning.
Data is recorded only if you have chosen to when you created CEC model object.
}
\keyword{datasets}

