% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{log.iters.cec}
\alias{log.iters.cec}
\title{log.iters.cec}
\format{\preformatted{ NULL
}}
\usage{
log.iters.cec
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print how many iterations it took to learn CEC model
}
\keyword{datasets}

