% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{methods}
\name{log.ncluster.cec}
\alias{log.ncluster.cec}
\title{log.ncluster.cec}
\usage{
\method{log}{ncluster.cec}(c)
}
\arguments{
\item{c}{CEC model object.}
}
\description{
Print number of clusters that has been recorded at each stage of learning.
Data is recorded only if you have chosen to when you created CEC model object.
}

