% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{logClusters}
\alias{logClusters}
\alias{logClusters,Rcpp_CecModel-method}
\title{logClusters}
\format{\preformatted{ NULL
}}
\usage{
logClusters(c)
}
\arguments{
\item{c}{object Trained CEC model object.}
}
\description{
Print number of clusters that has been recorded at each stage of learning.
Data is recorded only if you have chosen to when you created CEC model object.
}
\examples{
\dontrun{
logClusters(c)
}
}
\keyword{datasets}

