% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{pause}
\alias{pause}
\title{pause}
\usage{
pause(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Pause algorithm
}
\examples{
gng <- GNG(gng.preset.sphere(100))
pause(gng)
print(isRunning(gng))
}

