% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/svm.R
\docType{data}
\name{plot.svm}
\alias{plot.svm}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
plot(svm, pca=TRUE, log="x")
}
\arguments{
\item{pca}{Bool, use Principal Component Analysis for plot}

\item{dim1}{(optional) Dimension of dataset to plot on x axis, by default 1}

\item{dim2}{(optional) Dimension of dataset to plot on y axis, by default 2}

\item{log}{(optional) Use logarthic transformation for an axis, "x", "y", or "xy"}
}
\description{
Draws choosen dimenstions from a dataset on 2D plot coloring by class. NOTE: This function will be change to a package default function.
}
\keyword{datasets}

