% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.cec}
\alias{plot.cec}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
plot(cec)
}
\arguments{
\item{x}{CEC model object}

\item{slice}{list of dimentions chosen for display since plot is 2d}

\item{ellipses}{outline clusters}

\item{centers}{mark center of every cluster}
}
\description{
Plot clustering found
}
\details{
Plot CEC
}
\examples{
plot(cec, ellipses=TRUE, centers=FALSE)
}
\keyword{datasets}

