\docType{plot}
\name{plot.cec}
\alias{plot.cec}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
plot(cec)
}
\arguments{
  \item{x}{CEC model object.}

  \item{slice}{List of dimentions chosen for display since
  plot is 2D.}

  \item{ellipses}{Outline clusters.}

  \item{centers}{Marks center of every cluster.}
}
\description{
Plot clustering found on 2D plot coloring by cluster.
}
\details{
Plot CEC
}
\examples{
plot(cec)
}
\keyword{datasets}

