% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.plot.R
\docType{data}
\name{plot.cec}
\alias{plot.cec}
\title{plot}
\format{\preformatted{ NULL
}}
\usage{
plot(cec)
}
\arguments{
\item{x}{CEC model object.}

\item{slice}{List of dimentions chosen for display since plot is 2D.}

\item{ellipses}{Outline clusters.}

\item{centers}{Marks center of every cluster.}
}
\description{
Plot clustering found on 2D plot coloring by cluster.
}
\details{
Plot CEC
}
\examples{
plot(cec)
plot(cec, slice=c(1,3), ellipses=TRUE)
plot(cec, slice=c(1,2,3))
plot(cec, ellipses=TRUE, centers=FALSE)
}
\keyword{datasets}

