% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\name{plot.Rcpp_GNGServer}
\alias{plot.Rcpp_GNGServer}
\title{plot GNG object}
\usage{
\method{plot}{Rcpp_GNGServer}(x, vertex.color = gng.plot.color.cluster,
  layout = gng.plot.layout.v2d, mode = gng.plot.2d, vertex.size = 3, ...)
}
\arguments{
\item{x}{GNG object}

\item{vertex.color}{How to color vertexes. Possible values: \code{gng.plot.color.cluster} (vertex color is set to fastgreedy.community clustering),
\code{gng.plot.color.label} (rounds to integer label if present), \code{list of integers} (colors vertices according to provided list), \code{gng.plot.color.none} (every node is white),}

\item{layout}{Layout to be used when plotting. Possible values: \code{gng.plot.layour.igraph.v2d} (first two dimensions),
\code{gng.plot.layout.igraph.auto} (auto layout from igraph)  \code{gng.plot.layout.igraph.fruchterman.fast} (fast fruchterman reingold layout),or any function accepting igraph graph and returning layout}

\item{mode}{\code{gng.plot.2d} (igraph plot)
\code{gng.plot.2d.errors} (igraph plot with mean error log plot)}

\item{vertex.size}{Size of plotted vertices}

\item{...}{other arguments not used by this method.}
}
\description{
Plot resulting graph using igraph plotting
}
\details{
Plot GNG
}
\note{
If you want to "power-use" plotting and plot for instance a subgraph, you might be interested in
exporting igraph with convertToIGraph function
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
# Plots igraph using first 2 coordinates and colors according to clusters
plot(gng, mode=gng.plot.2d.errors, layout=gng.plot.layout.v2d, vertex.color=gng.plot.color.cluster)

# For more possibilities see gng.plot.* constants
}
}

