% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{predictComponent}
\alias{predictComponent}
\title{predictComponent}
\usage{
predictComponent(gng, c(1,1,1))
}
\arguments{
\item{x}{Can be either vector or data.frame.}
}
\description{
Finds connected component closest to given vector(s).
}
\details{
Find closest component
}
\examples{
# Find closest component to c(1,1,1)
predictComponent(gng,  c(1,1,1))
}

