% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.print.R
\docType{methods}
\name{print.Rcpp_CecModel}
\alias{print.Rcpp_CecModel}
\title{print}
\usage{
\method{print}{Rcpp_CecModel}(x, ...)
}
\arguments{
\item{x}{CEC object model.}

\item{...}{other arguments not used by this method.}
}
\description{
Print basic information about clusters found.
Presents a structure of the cec results object (clusters found)
}
\details{
Print CEC
}

