\docType{methods}
\name{print.cec}
\alias{print.cec}
\title{print}
\format{\preformatted{ NULL
}}
\usage{
print(cec)
}
\arguments{
  \item{x}{CEC object model.}
}
\description{
Print basic information about clusters found. Presents a
structure of the cec results object (clusters found)
}
\details{
Print CEC
}
\keyword{datasets}

