% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.print.R
\docType{methods}
\name{print.cec}
\alias{print.cec}
\title{print}
\usage{
print(cec)
}
\arguments{
\item{x}{CEC object model.}
}
\description{
Print basic information about clusters found.
Presents a structure of the cec results object (clusters found)
}
\details{
Print CEC
}

