% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\name{print.Rcpp_SVMClient}
\alias{print.Rcpp_SVMClient}
\title{Print SVM object}
\usage{
\method{print}{Rcpp_SVMClient}(x, ...)
}
\arguments{
\item{x}{Trained SVM object}

\item{...}{other arguments not used by this method.}
}
\description{
Prints short summary of a trained model.
}

