% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{class}
\name{Rcpp_CecConfiguration-class}
\alias{Rcpp_CecConfiguration-class}
\title{Class Rcpp_CecConfiguration.}
\description{
Class \code{Rcpp_CecConfiguration} defines a CEC model configuration class.
}
\section{Methods}{

\describe{
\item{\code{setAlgorithm(...)}}{void setAlgorithm(std::string)}

\item{\code{setCentroids(...)}}{void setCentroids(Rcpp::List)}

\item{\code{setCov(...)}}{void setCov(Rcpp::Matrix<14, Rcpp::PreserveStorage>)}

\item{\code{setDataSet(...)}}{void setDataSet(Rcpp::Matrix<14, Rcpp::PreserveStorage>)}

\item{\code{setEps(...)}}{void setEps(double)}

\item{\code{setFunction(...)}}{void setFunction(Rcpp::Function_Impl<Rcpp::PreserveStorage>)}

\item{\code{setItmax(...)}}{void setItmax(int)}

\item{\code{setLogCluster(...)}}{void setLogCluster(bool)}

\item{\code{setLogEnergy(...)}}{void setLogEnergy(bool)}

\item{\code{setMethodInit(...)}}{void setMethodInit(std::string)}

\item{\code{setMethodType(...)}}{void setMethodType(std::string)}

\item{\code{setMix(...)}}{void setMix(Rcpp::List)}

\item{\code{setNrOfClusters(...)}}{void setNrOfClusters(unsigned int)}

\item{\code{setNstart(...)}}{void setNstart(unsigned int)}

\item{\code{setR(...)}}{void setR(double)}

\item{\code{setSeed(...)}}{void setSeed(int)}
}}

