% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{class}
\name{Rcpp_CecModel-class}
\alias{Rcpp_CecModel-class}
\title{Class Rcpp_CecModel.}
\description{
Class \code{Rcpp_CecModel} defines a CEC model class.
}
\section{Methods}{

\describe{
\item{\code{centers()}}{std::__1::vector<arma::Row<double>, std::__1::allocator<arma::Row<double> > > centers()  const }

\item{\code{clustering()}}{std::__1::vector<unsigned int, std::__1::allocator<unsigned int> > clustering()  const }

\item{\code{covMatrix()}}{std::__1::vector<arma::Mat<double>, std::__1::allocator<arma::Mat<double> > > covMatrix()  const }

\item{\code{energy()}}{double energy()  const }

\item{\code{getDataset()}}{arma::Mat<double> getDataset()}

\item{\code{log.energy()}}{std::__1::list<double, std::__1::allocator<double> > log.energy()  const }

\item{\code{log.iters()}}{unsigned int log.iters()  const }

\item{\code{log.ncluster()}}{std::__1::list<unsigned int, std::__1::allocator<unsigned int> > log.ncluster()  const }

\item{\code{predict(...)}}{unsigned int predict(std::__1::vector<double, std::__1::allocator<double> >)}

\item{\code{runAll()}}{void runAll()}

\item{\code{runOneIteration()}}{void runOneIteration()}
}}

