% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\name{runAll}
\alias{runAll}
\alias{runAll,Rcpp_CecModel-method}
\title{runAll}
\usage{
runAll(c)
}
\arguments{
\item{c}{Trained CEC model object.}
}
\description{
Starts whole algorithm again with same parameters
}

