% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.R
\docType{data}
\name{runOneIteration}
\alias{runOneIteration}
\alias{runOneIteration,Rcpp_CecModel-method}
\title{runOneIteration}
\format{\preformatted{ NULL
}}
\usage{
runOneIteration
}
\description{
runs one iteration of algorithm
}
\keyword{datasets}

