% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cec.summary.R
\docType{methods}
\name{summary.Rcpp_CecModel}
\alias{summary.Rcpp_CecModel}
\title{summary}
\usage{
\method{summary}{Rcpp_CecModel}(object, ...)
}
\arguments{
\item{object}{CEC model object.}

\item{...}{other arguments not used by this method.}
}
\description{
Print detailed information about CEC model object
}
\details{
Summary CEC
}

