% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cec.summary.R
\docType{methods}
\name{summary.cec}
\alias{summary.cec}
\title{summary}
\usage{
summary(cec)
}
\arguments{
\item{object}{CEC model object.}
}
\description{
Print detailed information about CEC model object
}
\details{
Summary CEC
}

