% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict.svm}
\alias{predict.svm}
\alias{test}
\title{Predict}
\usage{
predict(svm, x)
}
\arguments{
\item{object}{Trained SVM object.}

\item{x}{unlabeled data, note that each entry needs to be the same dimentionality as training examples.}
}
\description{
Returns predicted classes for provided test examples.
}

