% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/svm.R
\docType{methods}
\name{predict.svm.gmum}
\alias{predict.svm.gmum}
\alias{test}
\title{Predict}
\usage{
predict(svm, x)
}
\arguments{
\item{object}{Trained SVM object.}

\item{x}{unlabeled data, note that each entry needs to be the same dimentionality as training examples.}
}
\description{
Returns predicted classes for provided test examples.
}

