% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SVM}
\alias{SVM}
\title{SVM}
\format{\preformatted{ NULL
}}
\usage{
SVM
}
\arguments{
\item{x}{Dataset without labels}

\item{y}{Labels}

\item{lib}{Desired SVM Library, avialable are: libsvm}

\item{kernel}{Kernel type, avialable are: linear, poly, rbf, sigmoid}

\item{prep}{Preprocess method, avialable are: none, 2e}

\item{mclass}{Multiclass wariant, avialable are: none}

\item{C}{Cost/Complexity parameter}

\item{gamma}{Gamma parameter for poly, rbf and sigmoid kernels}

\item{coef0}{Coef0 for poly and sigmoid kernels}

\item{degree}{Degree for poly kernel}

\item{shrinking}{Whether to use shrinking heuristics}

\item{probability}{Whether to train a model for probability estimates}

\item{cache_size}{Cache size}

\item{tol}{Tolerance of termination criterion}
}
\value{
SVM model object
}
\description{
Create SVM model object. If any parameter will be omitted a default value will be used
}
\examples{
svm <- SVM(lib = "libsvm", kernel = "linear", C = 1, gamma = 0.01, coef0 = 0,degree = 3)
}
\keyword{datasets}

