% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gng.R
\docType{methods}
\name{terminate}
\alias{terminate}
\title{terminate}
\usage{
terminate(object)
}
\arguments{
\item{object}{GNG object}
}
\description{
Terminate algorithm
}
\examples{
\dontrun{
gng <- GNG(scaled.wine)
terminate(gng)
}
}

